day sunday
now its been a whole week of working on css, over the break I have completed the homepage, checkout, the about, and some more maybe. we need to improve some of it but the core parts are complete. There are notes on each css code that I don't understand or need help with. Sadly I never wanted to learn java so I didn't learn it. still have around 4 more pages left.